Refactor(#5) : DTO상의 average_cost Long 타입으로 변경

+++ src/main/java/cowing/auth/dto/PortfolioDto.java
@@ -9,5 +9,5 @@ import java.math.BigDecimal;
 public record PortfolioDto(
         @JsonProperty("name") @NotBlank String name,
         @JsonProperty("quantity")  @NotNull BigDecimal quantity,
-        @JsonProperty("average_cost")  @NotNull Integer averageCost) {
+        @JsonProperty("average_cost")  @NotNull Long averageCost) {
 }
